Пофиксил тесты + по хуйне

+++ src/main/java/ru/readysetcock/fate_telegram_bot/repository/TaroCardRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.readysetcock.fate_telegram_bot.model.domain.TaroCardEntity;
 
 @Repository
-public interface TaroCardRepository extends CrudRepository<TaroCardEntity, Long> {
+public interface TaroCardRepository extends CrudRepository<TaroCardEntity, Integer> {
 }
